refactor(api): move copy target profile route and controller to subdomain
