+++ api/src/prescription/target-profile/application/admin-target-profile-controller.js
@@ -1,8 +1,10 @@
 import dayjs from 'dayjs';
 
+import { usecases } from '../../../../lib/domain/usecases/index.js';
 import * as organizationSerializer from '../../../organizational-entities/infrastructure/serializers/jsonapi/organization-serializer.js';
+import { withTransaction } from '../../../shared/domain/DomainTransaction.js';
 import { escapeFileName } from '../../../shared/infrastructure/utils/request-response-utils.js';
-import { usecases } from '../domain/usecases/index.js';
+import { usecases as prescriptionTargetProfileUsecases } from '../domain/usecases/index.js';
 import * as targetProfileAttachOrganizationSerializer from '../infrastructure/serializers/jsonapi/target-profile-attach-organization-serializer.js';
 import * as targetProfileDetachOrganizationsSerializer from '../infrastructure/serializers/jsonapi/target-profile-detach-organizations-serializer.js';
 import * as targetProfileSerializer from '../infrastructure/serializers/jsonapi/target-profile-serializer.js';
@@ -11,7 +13,9 @@ import * as learningContentPDFPresenter from './presenter/pdf/learning-content-p
 const getContentAsJsonFile = async function (request, h) {
   const targetProfileId = request.params.id;
 
-  const { jsonContent, targetProfileName } = await usecases.getTargetProfileContentAsJson({ targetProfileId });
+  const { jsonContent, targetProfileName } = await prescriptionTargetProfileUsecases.getTargetProfileContentAsJson({
+    targetProfileId,
+  });
 
   const filename = escapeFileName(`${dayjs().format('YYYYMMDD')}_profil_cible_${targetProfileName}`);
 
@@ -24,7 +28,10 @@ const getContentAsJsonFile = async function (request, h) {
 const attachOrganizations = async function (request, h, dependencies = { targetProfileAttachOrganizationSerializer }) {
   const organizationIds = request.payload['organization-ids'];
   const targetProfileId = request.params.id;
-  const results = await usecases.attachOrganizationsToTargetProfile({ targetProfileId, organizationIds });
+  const results = await prescriptionTargetProfileUsecases.attachOrganizationsToTargetProfile({
+    targetProfileId,
+    organizationIds,
+  });
 
   return h
     .response(dependencies.targetProfileAttachOrganizationSerializer.serialize({ ...results, targetProfileId }))
@@ -34,7 +41,10 @@ const attachOrganizations = async function (request, h, dependencies = { targetP
 const attachOrganizationsFromExistingTargetProfile = async function (request, h) {
   const existingTargetProfileId = request.payload['target-profile-id'];
   const targetProfileId = request.params.id;
-  await usecases.attachOrganizationsFromExistingTargetProfile({ targetProfileId, existingTargetProfileId });
+  await prescriptionTargetProfileUsecases.attachOrganizationsFromExistingTargetProfile({
+    targetProfileId,
+    existingTargetProfileId,
+  });
   return h.response({}).code(204);
 };
 
@@ -42,10 +52,11 @@ const getLearningContentAsPdf = async function (request, h, dependencies = { lea
   const targetProfileId = request.params.id;
   const { language } = request.query;
 
-  const { learningContent, targetProfileName } = await usecases.getLearningContentByTargetProfile({
-    targetProfileId,
-    language,
-  });
+  const { learningContent, targetProfileName } =
+    await prescriptionTargetProfileUsecases.getLearningContentByTargetProfile({
+      targetProfileId,
+      language,
+    });
 
   const filename = escapeFileName(`${dayjs().format('YYYYMMDD')}_profil_cible_${targetProfileName}`);
 
@@ -64,7 +75,7 @@ const getLearningContentAsPdf = async function (request, h, dependencies = { lea
 const attachTargetProfiles = async function (request, h) {
   const targetProfileIds = request.payload['target-profile-ids'];
   const organizationId = request.params.organizationId;
-  await usecases.attachTargetProfilesToOrganization({ organizationId, targetProfileIds });
+  await prescriptionTargetProfileUsecases.attachTargetProfilesToOrganization({ organizationId, targetProfileIds });
 
   return h.response({}).code(204);
 };
@@ -73,7 +84,7 @@ const detachOrganizations = async function (request, h, dependencies = { targetP
   const { organizationIds } = request.deserializedPayload;
   const targetProfileId = request.params.targetProfileId;
 
-  const detachedOrganizationIds = await usecases.detachOrganizationsFromTargetProfile({
+  const detachedOrganizationIds = await prescriptionTargetProfileUsecases.detachOrganizationsFromTargetProfile({
     targetProfileId,
     organizationIds,
   });
@@ -88,14 +99,14 @@ const detachOrganizations = async function (request, h, dependencies = { targetP
 const outdateTargetProfile = async function (request, h) {
   const id = request.params.targetProfileId;
 
-  await usecases.outdateTargetProfile({ id });
+  await prescriptionTargetProfileUsecases.outdateTargetProfile({ id });
   return h.response({}).code(204);
 };
 
 const markTargetProfileAsSimplifiedAccess = async function (request, h) {
   const id = request.params.targetProfileId;
 
-  const targetProfile = await usecases.markTargetProfileAsSimplifiedAccess({ id });
+  const targetProfile = await prescriptionTargetProfileUsecases.markTargetProfileAsSimplifiedAccess({ id });
   return h.response(targetProfileSerializer.serialize(targetProfile));
 };
 
@@ -103,13 +114,33 @@ const findPaginatedFilteredTargetProfileOrganizations = async function (request)
   const targetProfileId = request.params.targetProfileId;
   const { filter, page } = request.query;
 
-  const { models: organizations, pagination } = await usecases.findPaginatedFilteredOrganizationByTargetProfileId({
-    targetProfileId,
-    filter,
-    page,
-  });
+  const { models: organizations, pagination } =
+    await prescriptionTargetProfileUsecases.findPaginatedFilteredOrganizationByTargetProfileId({
+      targetProfileId,
+      filter,
+      page,
+    });
   return organizationSerializer.serialize(organizations, pagination);
 };
+
+const copyTargetProfile = withTransaction(async (request) => {
+  const targetProfileIdToCopy = request.params.targetProfileId;
+  const copiedTargetProfileId = await prescriptionTargetProfileUsecases.copyTargetProfile({
+    targetProfileId: targetProfileIdToCopy,
+  });
+  await Promise.all([
+    await usecases.copyTargetProfileBadges({
+      originTargetProfileId: targetProfileIdToCopy,
+      destinationTargetProfileId: copiedTargetProfileId,
+    }),
+    await usecases.copyTargetProfileStages({
+      originTargetProfileId: targetProfileIdToCopy,
+      destinationTargetProfileId: copiedTargetProfileId,
+    }),
+  ]);
+
+  return copiedTargetProfileId;
+});
 
 const targetProfileController = {
   outdateTargetProfile,
@@ -121,6 +152,7 @@ const targetProfileController = {
   getContentAsJsonFile,
   getLearningContentAsPdf,
   findPaginatedFilteredTargetProfileOrganizations,
+  copyTargetProfile,
 };
 
 export { targetProfileController };
