+++ api/tests/prescription/target-profile/acceptance/application/admin-target-profile-route_test.js
@@ -462,4 +462,79 @@ describe('Acceptance | TargetProfile | Application | Route | admin-target-profil
       expect(response.result.data[0].id).to.equal(organizationId.toString());
     });
   });
+
+  describe('POST /api/admin/target-profiles/{id}/copy', function () {
+    let user;
+
+    beforeEach(function () {
+      user = databaseBuilder.factory.buildUser.withRole();
+      const learningContentForTest = {
+        skills: [
+          {
+            id: 'recSkill1',
+            name: 'skill1',
+            status: 'actif',
+            tubeId: 'recTube1',
+          },
+        ],
+      };
+      mockLearningContent(learningContentForTest);
+      return databaseBuilder.commit();
+    });
+
+    it('should return the new target profile id and 200 status code', async function () {
+      // given
+      const tubeId = 'recArea1_Competence1_Tube1';
+      const targetProfile = databaseBuilder.factory.buildTargetProfile();
+      const targetProfileTube = databaseBuilder.factory.buildTargetProfileTube({
+        targetProfileId: targetProfile.id,
+        tubeId,
+        level: 1,
+      });
+
+      const badge = databaseBuilder.factory.buildBadge({ key: 'FIRST_BADGE', targetProfileId: targetProfile.id });
+      const badgeCriterion = databaseBuilder.factory.buildBadgeCriterion({ badgeId: badge.id });
+      const stage = databaseBuilder.factory.buildStage({ targetProfileId: targetProfile.id });
+      await databaseBuilder.commit();
+
+      // when
+      const response = await server.inject({
+        method: 'POST',
+        url: `/api/admin/target-profiles/${targetProfile.id}/copy`,
+        headers: { authorization: generateValidRequestAuthorizationHeader(user.id) },
+      });
+
+      // then
+      const { id: targetProfileId, name: targetProfileName } = await knex('target-profiles')
+        .where('id', response.result)
+        .first();
+      const {
+        id: targetProfileTubeId,
+        tubeId: copiedTubeId,
+        targetProfileId: tubeTargetProfileId,
+      } = await knex('target-profile_tubes').where({ targetProfileId }).first();
+
+      const { id: badgeId } = await knex('badges').where({ targetProfileId }).first();
+      const { id: badgeCriterionId, name: badgeCriterionName } = await knex('badge-criteria')
+        .where({ badgeId })
+        .first();
+      const { id: stageId, message } = await knex('stages').where({ targetProfileId }).first();
+
+      expect(response.statusCode).to.equal(200);
+
+      expect(targetProfileName).to.equal('[Copie] ' + targetProfile.name);
+      expect(targetProfileId).not.to.equal(targetProfile.id);
+
+      expect(targetProfileTubeId).not.to.equal(targetProfileTube.id);
+      expect(copiedTubeId).to.equal(targetProfileTube.tubeId);
+      expect(tubeTargetProfileId).to.equal(targetProfileId);
+
+      expect(badgeId).not.to.equal(badge.id);
+      expect(badgeCriterionId).not.to.equal(badgeCriterion.id);
+      expect(badgeCriterionName).to.equal(badgeCriterion.name);
+
+      expect(stageId).not.to.equal(stage.id);
+      expect(message).to.equal(stage.message);
+    });
+  });
 });
